+++ changelog.d/17595.misc
@@ -1,0 +1,1 @@
+Refactor sliding sync class into multiple files.
